feat: allow unused rest siblings
A common pattern for immutably removing fields from an
object is using rest properties, picking out the fields you _dont_
want and collecting the remaining fields into a new object.

This expands the config for `no-unused-vars` to allow unused
variables only when the are siblings of a rest property in
a destructuring assignment.

Semver: minor
Ref: LOG-7154