chore(test): Read fixtures with eslint rather than fs
eslint has a built-in way of reading source code from a file. Using
this style makes it a bit easier to add tests. There was also a missing
assertion for a required file that doesn't exist, and this kicked in
when re-organizing the test for eslint-plugin-logdna. It was fixed.

Semver: patch