feat(rules)!: switch to tap-no-deprecated-aliases rule
BREAKING CHANGE: Un-aliased tap assertion methods will now be enforced

tap@15.0.0 deprecated the use of aliases for assertion methods, which
invalidates the concept of a "preferred" alias. This switches over to
a new rule for requiring the use of un-aliased assertion methods.

This rule is auto-fixable, and the changes it enforces are backwards
compatible with older versions of tap.

Ref: LOG-9396